refactor : Change package name
* Change package name from result to gameresult
