+++ src/main/java/lotto/view/output/ResultView.java
@@ -2,9 +2,9 @@ package lotto.view.output;
 
 import lotto.model.Rank;
 import lotto.model.mylottos.MyLottos;
-import lotto.model.result.EarningRate;
-import lotto.model.result.GameResult;
-import lotto.model.result.MatchingResult;
+import lotto.model.gameresult.EarningRate;
+import lotto.model.gameresult.GameResult;
+import lotto.model.gameresult.MatchingResult;
 
 import static lotto.model.Rank.SECOND;
